test(action-bar): recategorize and clean up browser tests
Move tests that don't need a real browser from action-bar.test.browser.tsx
to action-bar.test.tsx (jsdom). Merge use-action-bar.test.ts into the new
jsdom file. Remove tests empirically verified to not contribute to coverage
(displayName, className, HTML tag).
